knotserver: git: fetch last commit time against current plumbing.Hash
This fixes the bug where paths wouldn't show up in non-default branches.

We were running 'git log' against the bare repository that was pointing
to the default branch, which would then return no output for that
specific path since -- understandably -- the path doesn't exist on the
default branch. This change simply passes the current g.h
(plumbing.Hash) to 'git log' which resolves the path in that ref.