nix: modules/spindle: Fix nix option to match spindle
There was a change the environment variables in spindle that were not
reflected in this nix module, specifically STEP_TIMEOUT was renamed to
WORKFLOW_TIMEOUT, which means that there was no way to set the timeout
in the nix module.

With the option name being updated, this is a breaking change that will
require you to update your nix config if you are using the spindle module.

Signed-off-by: Samuel Shuert <me@thecoded.prof>