appview: db: replace `Issue.IssueAt` to `Issue.Rkey`
Replace `Issue.IssueAt` to `Issue.Rkey`. `issue_at` column still exist,
but not used and will be removed in future.

Remove `db.SetIssueAt()` in flavor of `Rkey`.
Create rkey *before* adding to DB, and save same rkey to PDS.
This is safe as `tx` is will be dropped when `ctx` is canceled with
error.

Signed-off-by: Seongmin Lee <boltlessengineer@proton.me>