input.css: remove ss01 font-feature
tailwind seems to apply this to every html element. when applying
font-mono to an element, it inherits the ss01 font feature. this results
in a different `a` characters from when inside `pre` and `code` blocks,
where it resets the list of font-features. 

i can't seem to find a good fix for this issue. removing the ss01
feature altogether seems like a good fix, the number alternates in inter
aren't all that different in my opinion.

Signed-off-by: oppiliappan <me@oppi.li>