lexicons/git: fix lexicon schema for refUpdate
Broken down the nested object declarations into individual declarations since lexicon schemas forbid nested declarations, tested with atcute's lexicon codegen.

I changed the name of `pair` to `individualLanguageSize` as the previous name didn't make it obvious as for what the object was for, slight breaking change but it seems fine given that this lexicon schema wasn't working prior and that the object isn't being used in any unions.

Signed-off-by: mary.my.id <git@mary.my.id>