lexicons,api: define and generate lexicon queries for knotserver
sh.tangled.repo.{diff,compare} are ... pretty hard to port over thanks
to the old handlers directly returning really complex jsonified types.
gitdiff.File, gitdiff.PatchHeader etc. aren't super easy to port over
to lexicon schema definitions.
An interim solution is to simply retain those as is and NOT define a
schema, which is allowed by the spec[0].

> schema (object, optional): schema definition, either an object, a ref,
> or a union of refs. Used to describe JSON encoded responses, though
> schema is optional even for JSON responses.

lexgen doesn't adhere to the spec, and crashes out when you try to
specify an application/json output but don't define a schema. To work
around this, we set encoding to '*/*'.

[0]: https://atproto.com/specs/lexicon#query-and-procedure-http-api

Signed-off-by: Anirudh Oppiliappan <anirudh@tangled.sh>