appview/models: fix logic issue when converting label record -> label ops
the ordering of ops should be deletes first followed by additions, in
order for ingestions to be idempotent. after every optimistic update the
the DB, we ingest the same record via the firehose. an ordering
difference resulted in certain additions being marked as no-ops when in
reality, it should be an addition followed by a deletion.

Signed-off-by: oppiliappan <me@oppi.li>