appview/pages: swap all star buttons for same repository
On timeline, if there is multiple cards for same repository including a
star button, when user star/unstar the repository, only first `#starBtn`
is swapped with new button.

This can be fixed by adding `data-star-subject-at` attr to identify all
star buttons related to same repository and use `hx-swap-oob` to replace
all of them.

Signed-off-by: Seongmin Lee <git@boltless.me>